[form-factors] Move DKMR2017 parametrization to separate files
